✏️️fix: 타입 변경 및 estateSale 필드 추가

+++ wooreadmin/src/main/java/com/piehouse/wooreadmin/subscription/entity/Subscription.java
@@ -30,7 +30,7 @@ public class Subscription {
     private Long customer;
 
     @Column(nullable = false)
-    private Integer subTokenAmount;
+    private Long subTokenAmount;
 
     @CreationTimestamp
     @Column(updatable = false)
